+++ convolution.h
@@ -2,9 +2,9 @@
 #define MY_GLOBALS_H
 
 struct ImageData { 
-    int imgArray[132][132];
+    double imgArray[132][132];
 };
 
-void *conv(ImageData num, float[4][33][33]);
+void *conv(ImageData num, double[4][33][33]);
 
 #endif
