+++ gabor.h
@@ -3,13 +3,13 @@
 
 const int maxMap = 33;
 
-extern float (*map5)[maxMap][maxMap];
-extern float (*map9)[maxMap][maxMap];
-extern float (*map17)[maxMap][maxMap];
-extern float (*map33)[maxMap][maxMap];
-extern float (*map5i)[maxMap][maxMap];
-extern float (*map9i)[maxMap][maxMap];
-extern float (*map17i)[maxMap][maxMap];
-extern float (*map33i)[maxMap][maxMap];
+extern double (*map5)[maxMap][maxMap];
+extern double (*map9)[maxMap][maxMap];
+extern double (*map17)[maxMap][maxMap];
+extern double (*map33)[maxMap][maxMap];
+extern double (*map5i)[maxMap][maxMap];
+extern double (*map9i)[maxMap][maxMap];
+extern double (*map17i)[maxMap][maxMap];
+extern double (*map33i)[maxMap][maxMap];
 
 #endif
